fix: fix delete card

+++ controllers/movies.js
@@ -17,19 +17,27 @@ const createMovie = (req, res, next) => {
     .catch(next);
 };
 
+// const deleteMovie = (req, res, next) => {
+//   const { id } = req.params;
+//   Movie.findOne({ movieId: id, owner: req.user._id })
+//     .orFail(new NotFound(ERROR_MESSAGE.ERROR_FILM_NOT_FOUND))
+//     .then((movie) => {
+//       if (req.user._id === String(movie.owner)) {
+//         return Movie.deleteOne({ movieId: id })
+//           .then(() => res.status(200)
+//             .send({ message: `Фильм ${id} удален` }))
+//           .catch(next);
+//       }
+//       return next(new BadForbidden(ERROR_MESSAGE.ERROR_ACCESS_IS_DENIED));
+//     })
+//     .catch(next);
+// };
+
 const deleteMovie = (req, res, next) => {
   const { id } = req.params;
-  Movie.findOne({ movieId: id, owner: req.user._id })
-    .orFail(new NotFound(ERROR_MESSAGE.ERROR_FILM_NOT_FOUND))
-    .then((movie) => {
-      if (req.user._id === String(movie.owner)) {
-        return Movie.deleteOne({ movieId: id })
-          .then(() => res.status(200)
-            .send({ message: `Фильм ${id} удален` }))
-          .catch(next);
-      }
-      return next(new BadForbidden(ERROR_MESSAGE.ERROR_ACCESS_IS_DENIED));
-    })
+  return Movie.deleteOne({ movieId: id, owner: req.user._id })
+    .then(() => res.status(200)
+      .send({ message: `Фильм ${id} удален` }))
     .catch(next);
 };
